# Description
<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.
-->
Speed up Pipeline get and list.
* Add in memory LRU cache for Pipeline read and parsing.
* Optimize the configuration clean path method for dbt block
* Not repeatedly call get_repo_config in Pipeline LIST endpoint. Each get_repo_config takes about 3-4ms in my test.

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [x] Pipeline LIST endpoint: 170ms -> 50ms
- [x] Pipeline run LIST endpoint: 150ms -> 50ms


# Checklist
- [x] The PR is tagged with proper labels (bug, enhancement, feature, documentation)
- [x] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request -->